+++ src/LionWeb.Core/Core/Serialization/Delta/Command/DeltaProtocolPartitionCommandReceiver.cs
@@ -15,9 +15,8 @@
 // SPDX-FileCopyrightText: 2024 TRUMPF Laser SE and other contributors
 // SPDX-License-Identifier: Apache-2.0
 
-namespace LionWeb.Core.Serialization;
+namespace LionWeb.Core.Serialization.Delta.Command;
 
-using Delta.Command;
 using M1;
 using M1.Event.Partition;
 using M3;
