+++ test/LionWeb.Core.Test/Listener/EventTests.cs
@@ -18,7 +18,9 @@
 namespace LionWeb.Core.Test.Listener;
 
 using Core.Serialization;
+using Core.Serialization.Delta;
 using Core.Serialization.Delta.Command;
+using Core.Serialization.Delta.Event;
 using Core.Utilities;
 using Languages.Generated.V2024_1.Shapes.M2;
 using M1;
